backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
requests==2.31.0
voyageai==0.2.3
supabase==2.4.6
pypdf==4.2.0
python-dotenv==1.0.1
